Rename fileSettings property reset files to *.merger
